Search products

+++ src/main/java/fuse/appweb/shop/bill/BaseBillCalculator.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 public class BaseBillCalculator implements BillCalculator {
 
 	@Override
-	public int calculateBill(Order o,Map<String,Product> productsMap) {
+	public float calculateBill(Order o,Map<String,Product> productsMap) {
 		int total=0;
 		for (String p:o.getProducts()){
                     Product rp=productsMap.get(p);
